Move connector into baps3protocol, merge util with connector
