+++ baps3protocol/connector_test.go
@@ -1,4 +1,4 @@
-package util
+package baps3protocol
 
 import (
 	"testing"
